move decoder functions to the same package and simplify insertion function
